+++ pymc3/tests/test_minibatches.py
@@ -13,6 +13,35 @@ from pymc3.tests.helpers import select_by_precision
 from pymc3.theanof import GeneratorOp
 
 
+class _DataSampler(object):
+    """
+    Not for users
+    """
+    def __init__(self, data, batchsize=50, random_seed=42, dtype='floatX'):
+        self.dtype = theano.config.floatX if dtype == 'floatX' else dtype
+        self.rng = np.random.RandomState(random_seed)
+        self.data = data
+        self.n = batchsize
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        idx = (self.rng
+               .uniform(size=self.n,
+                        low=0.0,
+                        high=self.data.shape[0] - 1e-16)
+               .astype('int64'))
+        return np.asarray(self.data[idx], self.dtype)
+
+    next = __next__
+
+
+@pytest.fixture('module')
+def datagen():
+    return _DataSampler(np.random.uniform(size=(1000, 10)))
+
+
 def integers():
     i = 0
     while True:
